:art:style: 불필요한 코드 및 주석 제거

+++ src/main/java/com/opu/opube/feature/member/command/application/service/AuthService.java
@@ -94,7 +94,7 @@ public class AuthService {
         return TokenResponse.builder()
                 .accessToken(access)
                 .refreshToken(refresh)
-                .expiresInSeconds( (long) (jwtTokenProvider.getAccessExpirationSeconds()) )
+                .expiresInSeconds(jwtTokenProvider.getAccessExpirationSeconds())
                 .build();
     }
